ZylahMessagingAppBackend/MessagingApp/requirements.txt
channels
channels-redis
django
djangorestframework
djangorestframework-simplejwt
daphne
httpx
locust
pillow
psycopg2-binary
pytz
requests
stripe
